requirements.txt
git+https://github.com/YuanYuYuan/MIDP
tensorflow-gpu
tensorboardX
prettytable
torch
torchvision
opencv-python
pynrrd
ruamel.yaml
pathos
nibabel
pydicom
tqdm
psutil
json5
